feat: Refactor context providers and configs for Privy and Wagmi
Replaces individual Privy and Wagmi context files with a unified Providers component. Moves Privy and Wagmi configuration to dedicated files, updates supported chains, and removes obsolete config and context files. This improves modularity and centralizes provider setup.
