+++ context/providers.tsx
@@ -1,0 +1,24 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+
+import { PrivyProvider } from "@privy-io/react-auth";
+// Make sure to import these from `@privy-io/wagmi`, not `wagmi`
+import { WagmiProvider } from '@privy-io/wagmi';
+
+import { privyConfig } from "./privyConfig";
+import { wagmiConfig } from "./wagmiConfig";
+
+const queryClient = new QueryClient();
+
+export default function Providers({children}: {children: React.ReactNode}) {
+    return (
+        <PrivyProvider 
+            appId={process.env.NEXT_PUBLIC_PRIVY_APP_ID}
+            clientId={process.env.NEXT_PUBLIC_PRIVY_CLIENT_ID}
+            config={privyConfig}
+        >
+            <QueryClientProvider client={queryClient}>
+                <WagmiProvider config={wagmiConfig}>{children}</WagmiProvider>
+            </QueryClientProvider>
+        </PrivyProvider>
+    );
+}
