+++ context/wagmiConfig.ts
@@ -1,0 +1,17 @@
+import { createConfig, http } from "wagmi";
+import { celo, optimism } from "wagmi/chains";
+
+export const wagmiConfig = createConfig({
+  chains: [celo, optimism],
+  ssr: true,
+  transports: {
+    [celo.id]: http(process.env.NEXT_PUBLIC_ALCHEMY_RPC_URL),
+    [optimism.id]: http(),
+  },
+});
+
+declare module 'wagmi' {
+  interface Register {
+    config: typeof wagmiConfig
+  }
+}
